requirements.txt
django
djangorestframework







